Load UI components separate from Modeling languages
This allows us to use modeling languages without GTK dependency.
