+++ gaphor/RAAML/modelinglanguage.py
@@ -3,7 +3,6 @@ assets."""
 
 from typing import Iterable
 
-import gaphor.SysML.propertypages  # noqa
 from gaphor.abc import ModelingLanguage
 from gaphor.core import gettext
 from gaphor.diagram.diagramtoolbox import DiagramType, ToolboxDefinition
