+++ gaphor/SysML/modelinglanguage.py
@@ -4,7 +4,6 @@ assets."""
 from typing import Iterable
 
 import gaphor.SysML.drop  # noqa
-import gaphor.SysML.propertypages  # noqa
 from gaphor.abc import ModelingLanguage
 from gaphor.core import gettext
 from gaphor.diagram.diagramtoolbox import DiagramType, ToolboxDefinition
