+++ gaphor/SysML/propertypages.py
@@ -12,6 +12,11 @@ from gaphor.SysML.blocks.block import BlockItem
 from gaphor.SysML.requirements.requirement import RequirementItem
 from gaphor.UML.classes.classespropertypages import AttributesPage, OperationsPage
 
+
+def init():
+    pass
+
+
 new_builder = new_resource_builder("gaphor.SysML")
