+++ gaphor/UML/profiles/__init__.py
@@ -1,7 +1,5 @@
 from gaphor.UML.profiles import (
     extensionconnect,
-    metaclasspropertypage,
-    stereotypepropertypages,
 )
 from gaphor.UML.profiles.extension import ExtensionItem
 from gaphor.UML.profiles.packageimport import PackageImportItem
