+++ gaphor/UML/uicomponents.py
@@ -1,0 +1,24 @@
+# flake8: noqa F401
+
+from gaphor.UML.actions import (
+    actionseditors,
+    actionspropertypages,
+    activitymove,
+    activitypropertypage,
+    partitionpage,
+)
+from gaphor.UML.classes import (
+    associationpropertypages,
+    classeseditors,
+    classespropertypages,
+    dependencypropertypages,
+    enumerationpropertypages,
+)
+from gaphor.UML.deployments import deploymentpropertypage
+from gaphor.UML.interactions import interactionspropertypages
+from gaphor.UML.profiles import metaclasspropertypage, stereotypepropertypages
+from gaphor.UML.states import dropzone, propertypages
+
+
+def init():
+    """Allow this module to be loaded from an entrypoint."""
