+++ gaphor/diagram/general/uicomponents.py
@@ -1,0 +1,8 @@
+# flake8: noqa F401
+
+import gaphor.diagram.general.generaleditors
+import gaphor.diagram.general.generalpropertypages
+
+
+def init():
+    """Allows this module to be loaded from an entrypoint."""
